--- DNAStore/Executors/IExecutor.cs
@@ -1,9 +1,0 @@
-﻿namespace DNAStore.Executors;
-
-public interface IExecutor
-{
-    /// <summary>
-    /// Executes the request.
-    /// </summary>
-    void Run();
-}
