--- DNAStore/Executors/MinGCSkewLocation.cs
@@ -1,25 +1,0 @@
-using Bio.Sequence.Types;
-
-namespace DNAStore.Executors;
-
-internal class MinGCSkewLocation : BaseExecutor
-{
-    protected override void GetInputs()
-    {
-        Console.WriteLine("Enter Sequence To be analyzed");
-        sequence = new DNASequence(Console.ReadLine());
-    }
-
-    protected override void CalculateResult()
-    {
-        output = sequence.CalculateMinPrefixGCSkew();
-    }
-
-    protected override void OutputResult()
-    {
-        Console.WriteLine($"{string.Join(' ', output)}");
-    }
-
-    private NucleotideSequence? sequence;
-    private int[]? output;
-}
